Fixes for GPU (#57)
Initial support for running on GPUS
* register all MLIR passes and all dialects
* isolate ddpt and mlir features; hide all non-ddpt stuff in libs; move all ddpt internals into namespace DDPT
* use 32bit in examples and tests
* scalars become 0d-tensors of equivalent type in ewbinops

+++ examples/stencil-2d.py
@@ -120,8 +120,8 @@ def main():
     # there is certainly a more Pythonic way to initialize W,
     # but it will have no impact on performance.
     t0 = timer()
-    W = np.zeros(((2 * r + 1), (2 * r + 1)), dtype=np.float64)
-    B = np.zeros((n, n), dtype=np.float64)
+    W = np.zeros(((2 * r + 1), (2 * r + 1)), dtype=np.float32)
+    B = np.zeros((n, n), dtype=np.float32)
 
     if pattern == "star":
         stencil_size = 4 * r + 1
@@ -143,7 +143,7 @@ def main():
             W[r + j, r + j] = +1.0 / (4 * j * r)
             W[r - j, r - j] = -1.0 / (4 * j * r)
 
-    A = np.numpy.fromfunction(lambda i, j: i + j, (n, n), dtype=np.float64)
+    A = np.numpy.fromfunction(lambda i, j: i + j, (n, n), dtype=np.float32)
 
     for k in range(iterations + 1):
         # start timer after a warmup iteration
