+++ test/test_setget.py
@@ -9,7 +9,7 @@ class TestSetGet:
         for i in range(1, 5):
             N = (10**i) * 16
             print(N, N // 4, (N // 4) * 3)
-            doit = lambda aapi: aapi.ones((N, N), aapi.float64)[
+            doit = lambda aapi: aapi.ones((N, N), aapi.float32)[
                 0:N:2, N // 4 : (N // 4) * 3 : 2
             ]
 
@@ -17,16 +17,16 @@ class TestSetGet:
 
     def test_setitem1(self):
         def doit(aapi):
-            a = aapi.ones((16, 16), aapi.float64)
-            a[0:16:2, 0:16:2] = aapi.zeros([8, 8], aapi.float64)
+            a = aapi.ones((16, 16), aapi.float32)
+            a[0:16:2, 0:16:2] = aapi.zeros([8, 8], aapi.float32)
             return a
 
         assert runAndCompare(doit)
 
     def test_setitem2(self):
         def doit(aapi):
-            a = aapi.ones((16, 16), aapi.float64)
-            b = aapi.fromfunction(lambda i, j: 10 * i + j, (16, 16), dtype=aapi.float64)
+            a = aapi.ones((16, 16), aapi.float32)
+            b = aapi.fromfunction(lambda i, j: 10 * i + j, (16, 16), dtype=aapi.float32)
             a[1:8, 0:6] = b[0:7, 0:6]
             return a
 
@@ -34,8 +34,8 @@ class TestSetGet:
 
     def test_setitem3(self):
         def doit(aapi):
-            a = aapi.ones((16, 16), aapi.float64)
-            b = aapi.fromfunction(lambda i, j: 10 * i + j, (16, 16), dtype=aapi.float64)
+            a = aapi.ones((16, 16), aapi.float32)
+            b = aapi.fromfunction(lambda i, j: 10 * i + j, (16, 16), dtype=aapi.float32)
             a[7:16:3, 4:10:2] = b[4:7, 10:16:2]
             return a
 
@@ -44,8 +44,8 @@ class TestSetGet:
     def test_setitem4(self):
         # Note: test halo update without send buffer
         def doit(aapi):
-            a = aapi.ones((16, 16), aapi.float64)
-            b = aapi.fromfunction(lambda i, j: 10 * i + j, (16, 16), dtype=aapi.float64)
+            a = aapi.ones((16, 16), aapi.float32)
+            b = aapi.fromfunction(lambda i, j: 10 * i + j, (16, 16), dtype=aapi.float32)
             a[7:16:3, 0:16] = b[4:7, 0:16]
             return a
 
@@ -54,7 +54,7 @@ class TestSetGet:
     def test_setitem5(self):
         # Note: test assignment to one full local part
         def doit(aapi):
-            a = aapi.fromfunction(lambda i, j: 10 * i + j, (16, 16), dtype=aapi.int64)
+            a = aapi.fromfunction(lambda i, j: 10 * i + j, (16, 16), dtype=aapi.int32)
             a[0:10, 4:11] = a[0:10, 4:11]
             return a
 
@@ -63,8 +63,8 @@ class TestSetGet:
     def test_setitem6(self):
         def doit(aapi):
             n = 16
-            a = aapi.fromfunction(lambda i, j: i, (n, n), dtype=aapi.float64)
-            b = aapi.zeros((n + 1, n + 1), aapi.float64)
+            a = aapi.fromfunction(lambda i, j: i, (n, n), dtype=aapi.float32)
+            b = aapi.zeros((n + 1, n + 1), aapi.float32)
 
             b[1:n, 1:n] = a[1:n, 1:n]
             return b
@@ -75,8 +75,8 @@ class TestSetGet:
         # Note: assert halo does not segfault
         def doit(aapi):
             n = 1024
-            a = aapi.fromfunction(lambda i, j: i, (n, n), dtype=aapi.float64)
-            b = aapi.zeros((n, n), aapi.float64)
+            a = aapi.fromfunction(lambda i, j: i, (n, n), dtype=aapi.float32)
+            b = aapi.zeros((n, n), aapi.float32)
 
             b[1:n, 1:n] = a[1:n, 1:n]
             b[0, 1:n] = a[0, 1:n]
@@ -85,8 +85,8 @@ class TestSetGet:
         assert runAndCompare(doit)
 
     def test_colon(self):
-        a = dt.ones((16, 16), dt.float64)
-        b = dt.zeros((16, 16), dt.float64)
+        a = dt.ones((16, 16), dt.float32)
+        b = dt.zeros((16, 16), dt.float32)
         a[:, :] = b[:, :]
         r1 = dt.sum(a)
         assert float(r1) == 0
